refactor: general code improvements and fixes
- Remove redundant safe operators
- Remove unused code
- Remove unsused imports
- Fix use of deprecated URI constructor
- Reduce cognitive load for goToDefinition
- Merge definition with when statement
- Rename for consistency
- Remove redundant comments
