+++ server/src/main/kotlin/org/javacs/kt/KotlinTextDocumentService.kt
@@ -6,7 +6,7 @@ import org.eclipse.lsp4j.services.LanguageClient
 import org.eclipse.lsp4j.services.TextDocumentService
 import org.javacs.kt.codeaction.codeActions
 import org.javacs.kt.completion.completions
-import org.javacs.kt.definition.goToDefinition
+import org.javacs.kt.definition.DefinitionHandler
 import org.javacs.kt.diagnostic.convertDiagnostic
 import org.javacs.kt.formatting.FormattingService
 import org.javacs.kt.hover.hoverAt
@@ -36,36 +36,31 @@ import java.time.Duration
 import java.util.concurrent.CompletableFuture
 
 class KotlinTextDocumentService(
-    private val sf: SourceFiles,
-    private val sp: SourcePath,
+    private val sourceFiles: SourceFiles,
+    private val sourcePath: SourcePath,
     private val config: Configuration,
-    private val tempDirectory: TemporaryDirectory,
-    private val uriContentProvider: URIContentProvider,
-    private val cp: CompilerClassPath
+    tempDirectory: TemporaryDirectory,
+    uriContentProvider: URIContentProvider,
+    cp: CompilerClassPath
 ) : TextDocumentService, Closeable {
     private lateinit var client: LanguageClient
-    private val async = AsyncExecutor()
+    private val asyncExecutor = AsyncExecutor(name = "KotlinTextDocumentService")
     private val formattingService = FormattingService(config.formatting)
+    private val definitionHandler = DefinitionHandler(uriContentProvider.classContentProvider, tempDirectory, config.externalSources, cp)
 
-    var debounceLint = Debouncer(Duration.ofMillis(config.diagnostics.debounceTime))
+    var lintDebouncer = Debouncer(Duration.ofMillis(config.diagnostics.debounceTime))
     val lintTodo = mutableSetOf<URI>()
     var lintCount = 0
 
     var lintRecompilationCallback: () -> Unit
-        get() = sp.beforeCompileCallback
-        set(callback) { sp.beforeCompileCallback = callback }
-
-    private val TextDocumentItem.filePath: Path?
-        get() = parseURI(uri).filePath
+        get() = sourcePath.beforeCompileCallback
+        set(callback) { sourcePath.beforeCompileCallback = callback }
 
     private val TextDocumentIdentifier.filePath: Path?
         get() = parseURI(uri).filePath
 
-    private val TextDocumentIdentifier.isKotlinScript: Boolean
-        get() = uri.endsWith(".kts")
-
     private val TextDocumentIdentifier.content: String
-        get() = sp.content(parseURI(uri))
+        get() = sourcePath.content(parseURI(uri))
 
     fun connect(client: LanguageClient) {
         this.client = client
@@ -75,47 +70,47 @@ class KotlinTextDocumentService(
         ALWAYS, AFTER_DOT, NEVER
     }
 
-    private fun recover(position: TextDocumentPositionParams, recompile: Recompile): Pair<CompiledFile, Int>? {
+    private fun recover(position: TextDocumentPositionParams, recompile: Recompile = Recompile.NEVER): Pair<CompiledFile, Int>? {
         return recover(position.textDocument.uri, position.position, recompile)
     }
 
     private fun recover(uriString: String, position: Position, recompile: Recompile): Pair<CompiledFile, Int>? {
         val uri = parseURI(uriString)
-        if (!sf.isIncluded(uri)) {
+        if (!sourceFiles.isIncluded(uri)) {
             LOG.warn("URI is excluded, therefore cannot be recovered: $uri")
             return null
         }
-        val content = sp.content(uri)
+        val content = sourcePath.content(uri)
         val offset = offset(content, position.line, position.character)
         val shouldRecompile = when (recompile) {
             Recompile.ALWAYS -> true
             Recompile.AFTER_DOT -> offset > 0 && content[offset - 1] == '.'
             Recompile.NEVER -> false
         }
-        val compiled = if (shouldRecompile) sp.currentVersion(uri) else sp.latestCompiledVersion(uri)
+        val compiled = if (shouldRecompile) sourcePath.currentVersion(uri) else sourcePath.latestCompiledVersion(uri)
         return Pair(compiled, offset)
     }
 
-    override fun codeAction(params: CodeActionParams): CompletableFuture<List<Either<Command, CodeAction>>> = async.compute {
+    override fun codeAction(params: CodeActionParams): CompletableFuture<List<Either<Command, CodeAction>>> = asyncExecutor.compute {
         val (file, _) = recover(params.textDocument.uri, params.range.start, Recompile.NEVER) ?: return@compute emptyList()
-        codeActions(file, sp.index, params.range, params.context)
+        codeActions(file, sourcePath.index, params.range, params.context)
     }
 
-    override fun inlayHint(params: InlayHintParams): CompletableFuture<List<InlayHint>> = async.compute {
+    override fun inlayHint(params: InlayHintParams): CompletableFuture<List<InlayHint>> = asyncExecutor.compute {
         val (file, _) = recover(params.textDocument.uri, params.range.start, Recompile.ALWAYS) ?: return@compute emptyList()
         provideHints(file, config.inlayHints)
     }
 
-    override fun hover(position: HoverParams): CompletableFuture<Hover?> = async.compute {
+    override fun hover(position: HoverParams): CompletableFuture<Hover?> = asyncExecutor.compute {
         reportTime {
             LOG.info("Hovering at {}", describePosition(position))
 
-            val (file, cursor) = recover(position, Recompile.NEVER) ?: return@compute null
+            val (file, cursor) = recover(position) ?: return@compute null
             hoverAt(file, cursor) ?: noResult("No hover found at ${describePosition(position)}", null)
         }
     }
 
-    override fun documentHighlight(position: DocumentHighlightParams): CompletableFuture<List<DocumentHighlight>> = async.compute {
+    override fun documentHighlight(position: DocumentHighlightParams): CompletableFuture<List<DocumentHighlight>> = asyncExecutor.compute {
         val (file, cursor) = recover(position.textDocument.uri, position.position, Recompile.NEVER) ?: return@compute emptyList()
         documentHighlightsAt(file, cursor)
     }
@@ -124,19 +119,19 @@ class KotlinTextDocumentService(
         TODO("not implemented")
     }
 
-    override fun definition(position: DefinitionParams): CompletableFuture<Either<List<Location>, List<LocationLink>>> = async.compute {
+    override fun definition(position: DefinitionParams): CompletableFuture<Either<List<Location>, List<LocationLink>>> = asyncExecutor.compute {
         reportTime {
             LOG.info("Go-to-definition at {}", describePosition(position))
 
-            val (file, cursor) = recover(position, Recompile.NEVER) ?: return@compute Either.forLeft(emptyList())
-            goToDefinition(file, cursor, uriContentProvider.classContentProvider, tempDirectory, config.externalSources, cp)
+            val (file, cursor) = recover(position) ?: return@compute Either.forLeft(emptyList())
+            definitionHandler.goToDefinition(file, cursor)
                 ?.let(::listOf)
-                ?.let { Either.forLeft<List<Location>, List<LocationLink>>(it) }
+                ?.let { Either.forLeft(it) }
                 ?: noResult("Couldn't find definition at ${describePosition(position)}", Either.forLeft(emptyList()))
         }
     }
 
-    override fun rangeFormatting(params: DocumentRangeFormattingParams): CompletableFuture<List<TextEdit>> = async.compute {
+    override fun rangeFormatting(params: DocumentRangeFormattingParams): CompletableFuture<List<TextEdit>> = asyncExecutor.compute {
         val code = extractRange(params.textDocument.content, params.range)
         listOf(TextEdit(
             params.range,
@@ -148,17 +143,17 @@ class KotlinTextDocumentService(
         TODO("not implemented")
     }
 
-    override fun rename(params: RenameParams): CompletableFuture<WorkspaceEdit> = async.compute {
+    override fun rename(params: RenameParams): CompletableFuture<WorkspaceEdit?> = asyncExecutor.compute {
         val (file, cursor) = recover(params, Recompile.NEVER) ?: return@compute null
-        renameSymbol(file, cursor, sp, params.newName)
+        renameSymbol(file, cursor, sourcePath, params.newName)
     }
 
-    override fun completion(position: CompletionParams): CompletableFuture<Either<List<CompletionItem>, CompletionList>> = async.compute {
+    override fun completion(position: CompletionParams): CompletableFuture<Either<List<CompletionItem>, CompletionList>> = asyncExecutor.compute {
         reportTime {
             LOG.info("Completing at {}", describePosition(position))
 
-            val (file, cursor) = recover(position, Recompile.NEVER) ?: return@compute Either.forRight(CompletionList()) // TODO: Investigate when to recompile
-            val completions = completions(file, cursor, sp.index, config.completion)
+            val (file, cursor) = recover(position) ?: return@compute Either.forRight(CompletionList()) // TODO: Investigate when to recompile
+            val completions = completions(file, cursor, sourcePath.index, config.completion)
             LOG.info("Found {} items", completions.items.size)
 
             Either.forRight(completions)
@@ -169,13 +164,12 @@ class KotlinTextDocumentService(
         TODO("not implemented")
     }
 
-    @Suppress("DEPRECATION")
-    override fun documentSymbol(params: DocumentSymbolParams): CompletableFuture<List<Either<SymbolInformation, DocumentSymbol>>> = async.compute {
+    override fun documentSymbol(params: DocumentSymbolParams): CompletableFuture<List<Either<SymbolInformation, DocumentSymbol>>> = asyncExecutor.compute {
         LOG.info("Find symbols in {}", describeURI(params.textDocument.uri))
 
         reportTime {
             val uri = parseURI(params.textDocument.uri)
-            val parsed = sp.parsedFile(uri)
+            val parsed = sourcePath.parsedFile(uri)
 
             documentSymbols(parsed)
         }
@@ -183,7 +177,7 @@ class KotlinTextDocumentService(
 
     override fun didOpen(params: DidOpenTextDocumentParams) {
         val uri = parseURI(params.textDocument.uri)
-        sf.open(uri, params.textDocument.text, params.textDocument.version)
+        sourceFiles.open(uri, params.textDocument.text, params.textDocument.version)
         lintNow(uri)
     }
 
@@ -191,27 +185,27 @@ class KotlinTextDocumentService(
         // Lint after saving to prevent inconsistent diagnostics
         val uri = parseURI(params.textDocument.uri)
         lintNow(uri)
-        debounceLint.schedule {
-            sp.save(uri)
+        lintDebouncer.schedule {
+            sourcePath.save(uri)
         }
     }
 
-    override fun signatureHelp(position: SignatureHelpParams): CompletableFuture<SignatureHelp?> = async.compute {
+    override fun signatureHelp(position: SignatureHelpParams): CompletableFuture<SignatureHelp?> = asyncExecutor.compute {
         reportTime {
             LOG.info("Signature help at {}", describePosition(position))
 
-            val (file, cursor) = recover(position, Recompile.NEVER) ?: return@compute null
+            val (file, cursor) = recover(position) ?: return@compute null
             fetchSignatureHelpAt(file, cursor) ?: noResult("No function call around ${describePosition(position)}", null)
         }
     }
 
     override fun didClose(params: DidCloseTextDocumentParams) {
         val uri = parseURI(params.textDocument.uri)
-        sf.close(uri)
+        sourceFiles.close(uri)
         clearDiagnostics(uri)
     }
 
-    override fun formatting(params: DocumentFormattingParams): CompletableFuture<List<TextEdit>> = async.compute {
+    override fun formatting(params: DocumentFormattingParams): CompletableFuture<List<TextEdit>> = asyncExecutor.compute {
         val code = params.textDocument.content
         LOG.info("Formatting {}", describeURI(params.textDocument.uri))
         listOf(TextEdit(
@@ -222,25 +216,25 @@ class KotlinTextDocumentService(
 
     override fun didChange(params: DidChangeTextDocumentParams) {
         val uri = parseURI(params.textDocument.uri)
-        sf.edit(uri, params.textDocument.version, params.contentChanges)
+        sourceFiles.edit(uri, params.textDocument.version, params.contentChanges)
         lintLater(uri)
     }
 
-    override fun references(position: ReferenceParams) = async.compute {
+    override fun references(position: ReferenceParams) = asyncExecutor.compute {
         position.textDocument.filePath
             ?.let { file ->
-                val content = sp.content(parseURI(position.textDocument.uri))
+                val content = sourcePath.content(parseURI(position.textDocument.uri))
                 val offset = offset(content, position.position.line, position.position.character)
-                findReferences(file, offset, sp)
+                findReferences(file, offset, sourcePath)
             }
     }
 
-    override fun semanticTokensFull(params: SemanticTokensParams) = async.compute {
+    override fun semanticTokensFull(params: SemanticTokensParams) = asyncExecutor.compute {
         LOG.info("Full semantic tokens in {}", describeURI(params.textDocument.uri))
 
         reportTime {
             val uri = parseURI(params.textDocument.uri)
-            val file = sp.currentVersion(uri)
+            val file = sourcePath.currentVersion(uri)
 
             val tokens = encodedSemanticTokens(file)
             LOG.info("Found {} tokens", tokens.size)
@@ -249,12 +243,12 @@ class KotlinTextDocumentService(
         }
     }
 
-    override fun semanticTokensRange(params: SemanticTokensRangeParams) = async.compute {
+    override fun semanticTokensRange(params: SemanticTokensRangeParams) = asyncExecutor.compute {
         LOG.info("Ranged semantic tokens in {}", describeURI(params.textDocument.uri))
 
         reportTime {
             val uri = parseURI(params.textDocument.uri)
-            val file = sp.currentVersion(uri)
+            val file = sourcePath.currentVersion(uri)
 
             val tokens = encodedSemanticTokens(file, params.range)
             LOG.info("Found {} tokens", tokens.size)
@@ -271,15 +265,15 @@ class KotlinTextDocumentService(
         return "${describeURI(position.textDocument.uri)} ${position.position.line + 1}:${position.position.character + 1}"
     }
 
-    public fun updateDebouncer() {
-        debounceLint = Debouncer(Duration.ofMillis(config.diagnostics.debounceTime))
+    fun updateDebouncer() {
+        lintDebouncer = Debouncer(Duration.ofMillis(config.diagnostics.debounceTime))
     }
 
     fun lintAll() {
-        debounceLint.submitImmediately {
-            sp.compileAllFiles()
-            sp.saveAllFiles()
-            sp.refreshDependencyIndexes()
+        lintDebouncer.submitImmediately {
+            sourcePath.compileAllFiles()
+            sourcePath.saveAllFiles()
+            sourcePath.refreshDependencyIndexes()
         }
     }
 
@@ -291,18 +285,18 @@ class KotlinTextDocumentService(
 
     private fun lintLater(uri: URI) {
         lintTodo.add(uri)
-        debounceLint.schedule(::doLint)
+        lintDebouncer.schedule(::doLint)
     }
 
     private fun lintNow(file: URI) {
         lintTodo.add(file)
-        debounceLint.submitImmediately(::doLint)
+        lintDebouncer.submitImmediately(::doLint)
     }
 
     private fun doLint(cancelCallback: () -> Boolean) {
         LOG.info("Linting {}", describeURIs(lintTodo))
         val files = clearLint()
-        val context = sp.compileFiles(files)
+        val context = sourcePath.compileFiles(files)
         if (!cancelCallback.invoke()) {
             reportDiagnostics(files, context.diagnostics)
         }
@@ -316,7 +310,7 @@ class KotlinTextDocumentService(
         val byFile = langServerDiagnostics.groupBy({ it.first }, { it.second })
 
         for ((uri, diagnostics) in byFile) {
-            if (sf.isOpen(uri)) {
+            if (sourceFiles.isOpen(uri)) {
                 client.publishDiagnostics(PublishDiagnosticsParams(uri.toString(), diagnostics))
 
                 LOG.info("Reported {} diagnostics in {}", diagnostics.size, describeURI(uri))
@@ -338,13 +332,13 @@ class KotlinTextDocumentService(
         client.publishDiagnostics(PublishDiagnosticsParams(uri.toString(), listOf()))
     }
 
-    private fun shutdownExecutors(awaitTermination: Boolean) {
-        async.shutdown(awaitTermination)
-        debounceLint.shutdown(awaitTermination)
+    private fun shutdownExecutors(awaitTermination: Boolean = true) {
+        asyncExecutor.shutdown(awaitTermination)
+        lintDebouncer.shutdown(awaitTermination)
     }
 
     override fun close() {
-        shutdownExecutors(awaitTermination = true)
+        shutdownExecutors()
     }
 }
