+++ server/src/main/kotlin/org/javacs/kt/codeaction/quickfix/ImplementAbstractMembersQuickFix.kt
@@ -4,9 +4,6 @@ import org.eclipse.lsp4j.*
 import org.eclipse.lsp4j.jsonrpc.messages.Either
 import org.javacs.kt.CompiledFile
 import org.javacs.kt.index.SymbolIndex
-import org.javacs.kt.position.offset
-import org.javacs.kt.position.position
-import org.javacs.kt.util.toPath
 import org.javacs.kt.overridemembers.createFunctionStub
 import org.javacs.kt.overridemembers.createVariableStub
 import org.javacs.kt.overridemembers.getClassDescriptor
@@ -15,40 +12,31 @@ import org.javacs.kt.overridemembers.getNewMembersStartPosition
 import org.javacs.kt.overridemembers.getSuperClassTypeProjections
 import org.javacs.kt.overridemembers.hasNoBody
 import org.javacs.kt.overridemembers.overridesDeclaration
-import org.jetbrains.kotlin.descriptors.ClassDescriptor
-import org.jetbrains.kotlin.descriptors.ClassConstructorDescriptor
-import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
+import org.javacs.kt.position.offset
+import org.javacs.kt.util.toPath
 import org.jetbrains.kotlin.descriptors.FunctionDescriptor
+import org.jetbrains.kotlin.descriptors.Modality
 import org.jetbrains.kotlin.descriptors.PropertyDescriptor
 import org.jetbrains.kotlin.descriptors.isInterface
-import org.jetbrains.kotlin.descriptors.Modality
-import org.jetbrains.kotlin.js.resolve.diagnostics.findPsi
-import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.KtClass
-import org.jetbrains.kotlin.psi.KtDeclaration
-import org.jetbrains.kotlin.psi.KtNamedFunction
-import org.jetbrains.kotlin.psi.KtSimpleNameExpression
-import org.jetbrains.kotlin.psi.KtSuperTypeListEntry
-import org.jetbrains.kotlin.psi.KtTypeArgumentList
-import org.jetbrains.kotlin.psi.KtTypeReference
-import org.jetbrains.kotlin.psi.psiUtil.containingClass
-import org.jetbrains.kotlin.psi.psiUtil.endOffset
-import org.jetbrains.kotlin.psi.psiUtil.isAbstract
 import org.jetbrains.kotlin.psi.psiUtil.startOffset
 import org.jetbrains.kotlin.resolve.diagnostics.Diagnostics
-import org.jetbrains.kotlin.types.KotlinType
-import org.jetbrains.kotlin.types.TypeProjection
-import org.jetbrains.kotlin.types.typeUtil.asTypeProjection
 
 class ImplementAbstractMembersQuickFix : QuickFix {
-    override fun compute(file: CompiledFile, index: SymbolIndex, range: Range, diagnostics: List<Diagnostic>): List<Either<Command, CodeAction>> {
+    override fun compute(
+        file: CompiledFile,
+        index: SymbolIndex,
+        range: Range,
+        diagnostics: List<Diagnostic>,
+    ): List<Either<Command, CodeAction>> {
         val diagnostic = findDiagnosticMatch(diagnostics, range)
 
         val startCursor = offset(file.content, range.start)
         val endCursor = offset(file.content, range.end)
         val kotlinDiagnostics = file.compile.diagnostics
-        
-        // If the client side and the server side diagnostics contain a valid diagnostic for this range.
+
+        // If the client side and the server side diagnostics contain a valid diagnostic for this
+        // range.
         if (diagnostic != null && anyDiagnosticMatch(kotlinDiagnostics, startCursor, endCursor)) {
             // Get the class with the missing members
             val kotlinClass = file.parseAtPoint(startCursor)
@@ -62,14 +50,30 @@ class ImplementAbstractMembersQuickFix : QuickFix {
 
                 // Get the location where the new code will be placed
                 val newMembersStartPosition = getNewMembersStartPosition(file, kotlinClass)
-                val bodyAppendBeginning = listOf(TextEdit(Range(newMembersStartPosition, newMembersStartPosition), "{")).takeIf { kotlinClass.hasNoBody() } ?: emptyList()
-                val bodyAppendEnd = listOf(TextEdit(Range(newMembersStartPosition, newMembersStartPosition), System.lineSeparator() + "}")).takeIf { kotlinClass.hasNoBody() } ?: emptyList()
+                val bodyAppendBeginning =
+                    listOf(TextEdit(Range(newMembersStartPosition, newMembersStartPosition), "{"))
+                        .takeIf { kotlinClass.hasNoBody() } ?: emptyList()
+                val bodyAppendEnd =
+                    listOf(
+                            TextEdit(
+                                Range(newMembersStartPosition, newMembersStartPosition),
+                                System.lineSeparator() + "}",
+                            )
+                        )
+                        .takeIf { kotlinClass.hasNoBody() } ?: emptyList()
 
-                val textEdits = bodyAppendBeginning + membersToImplement.map {
-                    // We leave two new lines before the member is inserted
-                    val newText = System.lineSeparator() + System.lineSeparator() + padding + it
-                    TextEdit(Range(newMembersStartPosition, newMembersStartPosition), newText)
-                } + bodyAppendEnd
+                val textEdits =
+                    bodyAppendBeginning +
+                        membersToImplement.map {
+                            // We leave two new lines before the member is inserted
+                            val newText =
+                                System.lineSeparator() + System.lineSeparator() + padding + it
+                            TextEdit(
+                                Range(newMembersStartPosition, newMembersStartPosition),
+                                newText,
+                            )
+                        } +
+                        bodyAppendEnd
 
                 val codeAction = CodeAction()
                 codeAction.edit = WorkspaceEdit(mapOf(uri to textEdits))
@@ -81,36 +85,70 @@ class ImplementAbstractMembersQuickFix : QuickFix {
         }
         return listOf()
     }
-}
 
-fun findDiagnosticMatch(diagnostics: List<Diagnostic>, range: Range) =
-    diagnostics.find { diagnosticMatch(it, range, hashSetOf("ABSTRACT_MEMBER_NOT_IMPLEMENTED", "ABSTRACT_CLASS_MEMBER_NOT_IMPLEMENTED")) }
+    private fun findDiagnosticMatch(diagnostics: List<Diagnostic>, range: Range) =
+        diagnostics.find {
+            diagnosticMatch(
+                it,
+                range,
+                hashSetOf(
+                    "ABSTRACT_MEMBER_NOT_IMPLEMENTED",
+                    "ABSTRACT_CLASS_MEMBER_NOT_IMPLEMENTED",
+                ),
+            )
+        }
 
-private fun anyDiagnosticMatch(diagnostics: Diagnostics, startCursor: Int, endCursor: Int) =
-    diagnostics.any { diagnosticMatch(it, startCursor, endCursor, hashSetOf("ABSTRACT_MEMBER_NOT_IMPLEMENTED", "ABSTRACT_CLASS_MEMBER_NOT_IMPLEMENTED")) }
+    private fun anyDiagnosticMatch(diagnostics: Diagnostics, startCursor: Int, endCursor: Int) =
+        diagnostics.any {
+            diagnosticMatch(
+                it,
+                startCursor,
+                endCursor,
+                hashSetOf(
+                    "ABSTRACT_MEMBER_NOT_IMPLEMENTED",
+                    "ABSTRACT_CLASS_MEMBER_NOT_IMPLEMENTED",
+                ),
+            )
+        }
 
-private fun getAbstractMembersStubs(file: CompiledFile, kotlinClass: KtClass) =
-    // For each of the super types used by this class
-    kotlinClass.superTypeListEntries.mapNotNull {
-        // Find the definition of this super type
-        val referenceAtPoint = file.referenceExpressionAtPoint(it.startOffset)
-        val descriptor = referenceAtPoint?.second
+    private fun getAbstractMembersStubs(file: CompiledFile, kotlinClass: KtClass) =
+        // For each of the super types used by this class
+        kotlinClass.superTypeListEntries
+            .mapNotNull {
+                // Find the definition of this super type
+                val referenceAtPoint = file.referenceExpressionAtPoint(it.startOffset)
+                val descriptor = referenceAtPoint?.second
 
-        val classDescriptor = getClassDescriptor(descriptor)
-        
-        // If the super class is abstract or an interface
-        if (null != classDescriptor && (classDescriptor.kind.isInterface || classDescriptor.modality == Modality.ABSTRACT)) {
-            val superClassTypeArguments = getSuperClassTypeProjections(file, it)
-            classDescriptor.getMemberScope(superClassTypeArguments).getContributedDescriptors().filter { classMember ->
-               (classMember is FunctionDescriptor && classMember.modality == Modality.ABSTRACT && !overridesDeclaration(kotlinClass, classMember)) || (classMember is PropertyDescriptor && classMember.modality == Modality.ABSTRACT && !overridesDeclaration(kotlinClass, classMember))
-            }.mapNotNull { member ->
-                when (member) {
-                    is FunctionDescriptor -> createFunctionStub(member)
-                    is PropertyDescriptor -> createVariableStub(member)
-                    else -> null
+                val classDescriptor = getClassDescriptor(descriptor)
+
+                // If the super class is abstract or an interface
+                if (
+                    null != classDescriptor &&
+                        (classDescriptor.kind.isInterface ||
+                            classDescriptor.modality == Modality.ABSTRACT)
+                ) {
+                    val superClassTypeArguments = getSuperClassTypeProjections(file, it)
+                    classDescriptor
+                        .getMemberScope(superClassTypeArguments)
+                        .getContributedDescriptors()
+                        .filter { classMember ->
+                            (classMember is FunctionDescriptor &&
+                                classMember.modality == Modality.ABSTRACT &&
+                                !overridesDeclaration(kotlinClass, classMember)) ||
+                                (classMember is PropertyDescriptor &&
+                                    classMember.modality == Modality.ABSTRACT &&
+                                    !overridesDeclaration(kotlinClass, classMember))
+                        }
+                        .mapNotNull { member ->
+                            when (member) {
+                                is FunctionDescriptor -> createFunctionStub(member)
+                                is PropertyDescriptor -> createVariableStub(member)
+                                else -> null
+                            }
+                        }
+                } else {
+                    null
                 }
             }
-        } else {
-            null
-        }
-    }.flatten()
+            .flatten()
+}
