+++ server/src/main/kotlin/org/javacs/kt/codeaction/quickfix/QuickFix.kt
@@ -8,10 +8,9 @@ import org.eclipse.lsp4j.jsonrpc.messages.Either
 import org.javacs.kt.CompiledFile
 import org.javacs.kt.index.SymbolIndex
 import org.javacs.kt.util.isSubrangeOf
-import org.jetbrains.kotlin.resolve.diagnostics.Diagnostics
 import org.jetbrains.kotlin.diagnostics.Diagnostic as KotlinDiagnostic
 
-interface QuickFix {
+fun interface QuickFix {
     // Computes the quickfix. Return empty list if the quickfix is not valid or no alternatives exist.
     fun compute(file: CompiledFile, index: SymbolIndex, range: Range, diagnostics: List<Diagnostic>): List<Either<Command, CodeAction>>
 }
@@ -21,9 +20,3 @@ fun diagnosticMatch(diagnostic: Diagnostic, range: Range, diagnosticTypes: Set<S
 
 fun diagnosticMatch(diagnostic: KotlinDiagnostic, startCursor: Int, endCursor: Int, diagnosticTypes: Set<String>): Boolean =
     diagnostic.textRanges.any { it.startOffset <= startCursor && it.endOffset >= endCursor } && diagnosticTypes.contains(diagnostic.factory.name)
-
-fun findDiagnosticMatch(diagnostics: List<Diagnostic>, range: Range, diagnosticTypes: Set<String>) =
-    diagnostics.find { diagnosticMatch(it, range, diagnosticTypes) }
-
-fun anyDiagnosticMatch(diagnostics: Diagnostics, startCursor: Int, endCursor: Int, diagnosticTypes: Set<String>) =
-    diagnostics.any { diagnosticMatch(it, startCursor, endCursor, diagnosticTypes) }
