+++ server/src/main/kotlin/org/javacs/kt/docs/FindDoc.kt
@@ -11,9 +11,7 @@ import org.jetbrains.kotlin.psi.psiUtil.parents
 import org.jetbrains.kotlin.resolve.DescriptorToSourceUtils
 
 fun findDoc(declaration: DeclarationDescriptorWithSource): KDocTag? {
-    val source = DescriptorToSourceUtils.descriptorToDeclaration(declaration)?.navigationElement
-
-    return when (source) {
+    return when (val source = DescriptorToSourceUtils.descriptorToDeclaration(declaration)?.navigationElement) {
         is KtParameter -> {
             var container = source.parents.filterIsInstance<KtDeclaration>().firstOrNull() ?: return null
             if (container is KtPrimaryConstructor)
