+++ server/src/main/kotlin/org/javacs/kt/formatting/FormattingService.kt
@@ -1,8 +1,7 @@
 package org.javacs.kt.formatting
 
-import org.javacs.kt.Configuration
 import org.javacs.kt.FormattingConfiguration
-import org.eclipse.lsp4j.FormattingOptions as LspFromattingOptions
+import org.eclipse.lsp4j.FormattingOptions as LspFormattingOptions
 
 private const val DEFAULT_INDENT = 4
 
@@ -16,6 +15,6 @@ class FormattingService(private val config: FormattingConfiguration) {
 
     fun formatKotlinCode(
         code: String,
-        options: LspFromattingOptions = LspFromattingOptions(DEFAULT_INDENT, true)
+        options: LspFormattingOptions = LspFormattingOptions(DEFAULT_INDENT, true)
     ): String = this.formatter.format(code, options)
 }
