+++ server/src/main/kotlin/org/javacs/kt/j2k/JavaToKotlinConverter.kt
@@ -1,16 +1,12 @@
 package org.javacs.kt.j2k
 
 import com.intellij.lang.java.JavaLanguage
-import com.intellij.psi.PsiFileFactory
-import com.intellij.openapi.project.Project
-// import org.jetbrains.kotlin.j2k.JavaToKotlinTranslator
 import org.javacs.kt.LOG
 import org.javacs.kt.compiler.Compiler
-import org.javacs.kt.compiler.CompilationKind
-import org.javacs.kt.util.nonNull
+import org.javacs.kt.compiler.CompilationType
 
 fun convertJavaToKotlin(javaCode: String, compiler: Compiler): String {
-    val psiFactory = compiler.psiFileFactoryFor(CompilationKind.DEFAULT)
+    val psiFactory = compiler.psiFileFactoryFor(CompilationType.DEFAULT)
     val javaAST = psiFactory.createFileFromText("snippet.java", JavaLanguage.INSTANCE, javaCode)
     LOG.info("Parsed {} to {}", javaCode, javaAST)
