+++ server/src/main/kotlin/org/javacs/kt/progress/Progress.kt
@@ -5,28 +5,28 @@ import java.util.concurrent.CompletableFuture
 
 /** A facility for emitting progress notifications. */
 interface Progress : Closeable {
-    /**
-     * Updates the progress percentage. The
-     * value should be in the range [0, 100].
-     */
+    /** Updates the progress percentage. The value should be in the range [0, 100]. */
     fun update(message: String? = null, percent: Int? = null)
 
     object None : Progress {
-        override fun update(message: String?, percent: Int?) {}
+        override fun update(message: String?, percent: Int?) {
+            // Do nothing
+        }
 
-        override fun close() {}
+        override fun close() {
+            // Do nothing
+        }
     }
 
-    interface Factory {
-        /**
-         * Creates a new progress listener with
-         * the given label. The label is intended
-         * to be human-readable.
-         */
+    /**
+     * Creates a new progress listener with the given label. The label is intended to be
+     * human-readable.
+     */
+    fun interface Factory {
         fun create(label: String): CompletableFuture<Progress>
 
-        object None : Factory {
-            override fun create(label: String): CompletableFuture<Progress> = CompletableFuture.completedFuture(Progress.None)
+        companion object {
+            val None: Factory = Factory { CompletableFuture.completedFuture(Progress.None) }
         }
     }
 }
