+++ server/src/test/kotlin/org/javacs/kt/DebouncerTest.kt
@@ -7,17 +7,17 @@ import org.javacs.kt.util.Debouncer
 import java.time.Duration
 
 class DebouncerTest {
-    val debounce = Debouncer(Duration.ofSeconds(1))
+    val debouncer = Debouncer(Duration.ofSeconds(1))
     var counter = 0
 
     @Test fun callQuickly() {
         for (i in 1..10) {
-            debounce.schedule {
+            debouncer.schedule {
                 counter++
             }
         }
 
-        debounce.waitForPendingTask()
+        debouncer.waitForPendingTask()
 
         assertThat(counter, equalTo(1))
     }
