+++ shared/src/main/kotlin/org/javacs/kt/ScriptsConfiguration.kt
@@ -1,6 +1,6 @@
 package org.javacs.kt
 
-public data class ScriptsConfiguration(
+data class ScriptsConfiguration(
     /** Whether .kts scripts are handled. */
     var enabled: Boolean = false,
     /** Whether .gradle.kts scripts are handled. Only considered if scripts are enabled in general. */
