+++ shared/src/main/kotlin/org/javacs/kt/classpath/GradleClassPathResolver.kt
@@ -53,8 +53,8 @@ private fun gradleScriptToTempFile(scriptName: String, deleteOnExit: Boolean = f
     LOG.debug("Creating temporary gradle file {}", config.absolutePath)
 
     config.bufferedWriter().use { configWriter ->
-        GradleClassPathResolver::class.java.getResourceAsStream("/$scriptName").bufferedReader().use { configReader ->
-            configReader.copyTo(configWriter)
+        GradleClassPathResolver::class.java.getResourceAsStream("/$scriptName")?.bufferedReader().use { configReader ->
+            configReader?.copyTo(configWriter) ?: throw KotlinLSException("Could not find resource '$scriptName'")
         }
     }
 
@@ -66,23 +66,21 @@ private fun getGradleCommand(workspace: Path): Path {
     val wrapper = workspace.resolve(wrapperName).toAbsolutePath()
     if (Files.isExecutable(wrapper)) {
         return wrapper
-    } else {
-        return workspace.parent?.let(::getGradleCommand)
-            ?: findCommandOnPath("gradle")
-            ?: throw KotlinLSException("Could not find 'gradle' on PATH")
     }
+    return workspace.parent?.let(::getGradleCommand)
+        ?: findCommandOnPath("gradle")
+        ?: throw KotlinLSException("Could not find 'gradle' on PATH")
 }
 
 private fun readDependenciesViaGradleCLI(projectDirectory: Path, gradleScripts: List<String>, gradleTasks: List<String>): Set<Path> {
     LOG.info("Resolving dependencies for '{}' through Gradle's CLI using tasks {}...", projectDirectory.fileName, gradleTasks)
 
-    val tmpScripts = gradleScripts.map { gradleScriptToTempFile(it, deleteOnExit = false).toPath().toAbsolutePath() }
+    val tmpScripts = gradleScripts.map { gradleScriptToTempFile(it).toPath().toAbsolutePath() }
     val gradle = getGradleCommand(projectDirectory)
 
     val command = listOf(gradle.toString()) + tmpScripts.flatMap { listOf("-I", it.toString()) } + gradleTasks + listOf("--console=plain")
     val dependencies = findGradleCLIDependencies(command, projectDirectory)
-        ?.also { LOG.debug("Classpath for task {}", it) }
-        .orEmpty()
+        .also { LOG.debug("Classpath for task {}", it) }
         .filter { it.toString().lowercase().endsWith(".jar") || Files.isDirectory(it) } // Some Gradle plugins seem to cause this to output POMs, therefore filter JARs
         .toSet()
 
@@ -90,7 +88,7 @@ private fun readDependenciesViaGradleCLI(projectDirectory: Path, gradleScripts: 
     return dependencies
 }
 
-private fun findGradleCLIDependencies(command: List<String>, projectDirectory: Path): Set<Path>? {
+private fun findGradleCLIDependencies(command: List<String>, projectDirectory: Path): Set<Path> {
     val (result, errors) = execAndReadStdoutAndStderr(command, projectDirectory)
     if ("FAILURE: Build failed" in errors) {
         LOG.warn("Gradle task failed: {}", errors)
@@ -104,13 +102,11 @@ private fun findGradleCLIDependencies(command: List<String>, projectDirectory: P
     return parseGradleCLIDependencies(result)
 }
 
-private val artifactPattern by lazy { "kotlin-lsp-gradle (.+)(?:\r?\n)".toRegex() }
-private val gradleErrorWherePattern by lazy { "\\*\\s+Where:[\r\n]+(\\S\\.*)".toRegex() }
+private val artifactPattern by lazy { "kotlin-lsp-gradle (.+)\r?\n".toRegex() }
 
-private fun parseGradleCLIDependencies(output: String): Set<Path>? {
+private fun parseGradleCLIDependencies(output: String): Set<Path> {
     LOG.debug(output)
     val artifacts = artifactPattern.findAll(output)
-        .mapNotNull { Paths.get(it.groups[1]?.value) }
-        .filterNotNull()
+        .mapNotNull { it.groups[1]?.value?.let { it1 -> Paths.get(it1) } }
     return artifacts.toSet()
 }
