+++ shared/src/main/kotlin/org/javacs/kt/classpath/Home.kt
@@ -3,7 +3,6 @@ package org.javacs.kt.classpath
 import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.Paths
-import org.javacs.kt.util.KotlinLSException
 import org.javacs.kt.util.userHome
 
 private fun createPathOrNull(envVar: String): Path? = System.getenv(envVar)?.let(Paths::get)
@@ -11,8 +10,8 @@ private fun createPathOrNull(envVar: String): Path? = System.getenv(envVar)?.let
 private val possibleMavenRepositoryPaths =
     sequenceOf(
         createPathOrNull("MAVEN_REPOSITORY"),
-        createPathOrNull("MAVEN_HOME")?.let { it.resolve("repository") },
-        createPathOrNull("M2_HOME")?.let { it.resolve("repository") },
+        createPathOrNull("MAVEN_HOME")?.resolve("repository"),
+        createPathOrNull("M2_HOME")?.resolve("repository"),
         userHome.resolve(".m2/repository"),
     )
     .filterNotNull()
