+++ shared/src/main/kotlin/org/javacs/kt/util/Debouncer.kt
@@ -2,18 +2,16 @@ package org.javacs.kt.util
 
 import org.javacs.kt.LOG
 import java.time.Duration
-import java.util.function.Supplier
 import java.util.concurrent.atomic.AtomicReference
 import java.util.concurrent.ScheduledExecutorService
 import java.util.concurrent.Executors
 import java.util.concurrent.TimeUnit
-import java.util.concurrent.CompletableFuture
 import java.util.concurrent.Future
 
 private var threadCount = 0
 
 class Debouncer(
-    private val delay: Duration,
+    delay: Duration,
     private val executor: ScheduledExecutorService = Executors.newScheduledThreadPool(1) {
         Thread(it, "debounce${threadCount++}")
     }
@@ -24,7 +22,7 @@ class Debouncer(
     fun submitImmediately(task: (cancelCallback: () -> Boolean) -> Unit) {
         pendingTask?.cancel(false)
         val currentTaskRef = AtomicReference<Future<*>>()
-        val currentTask = executor.submit { task { currentTaskRef.get()?.isCancelled() ?: false } }
+        val currentTask = executor.submit { task { currentTaskRef.get()?.isCancelled ?: false } }
         currentTaskRef.set(currentTask)
         pendingTask = currentTask
     }
@@ -32,7 +30,7 @@ class Debouncer(
     fun schedule(task: (cancelCallback: () -> Boolean) -> Unit) {
         pendingTask?.cancel(false)
         val currentTaskRef = AtomicReference<Future<*>>()
-        val currentTask = executor.schedule({ task { currentTaskRef.get()?.isCancelled() ?: false } }, delayMs, TimeUnit.MILLISECONDS)
+        val currentTask = executor.schedule({ task { currentTaskRef.get()?.isCancelled ?: false } }, delayMs, TimeUnit.MILLISECONDS)
         currentTaskRef.set(currentTask)
         pendingTask = currentTask
     }
