+++ shared/src/main/kotlin/org/javacs/kt/util/KotlinLSException.kt
@@ -1,7 +1,4 @@
 package org.javacs.kt.util
 
-class KotlinLSException : RuntimeException {
-	constructor(msg: String) : super(msg) {}
-
-	constructor(msg: String, cause: Throwable) : super(msg, cause) {}
+class KotlinLSException(msg: String) : RuntimeException(msg) {
 }
