+++ shared/src/main/kotlin/org/javacs/kt/util/Utils.kt
@@ -4,13 +4,6 @@ import org.javacs.kt.LOG
 import java.io.PrintStream
 import java.nio.file.Path
 import java.nio.file.Paths
-import java.util.concurrent.CompletableFuture
-
-fun execAndReadStdout(shellCommand: List<String>, directory: Path): String {
-    val process = ProcessBuilder(shellCommand).directory(directory.toFile()).start()
-    val stdout = process.inputStream
-    return stdout.bufferedReader().use { it.readText() }
-}
 
 fun execAndReadStdoutAndStderr(shellCommand: List<String>, directory: Path): Pair<String, String> {
     val process = ProcessBuilder(shellCommand).directory(directory.toFile()).start()
@@ -34,14 +27,13 @@ inline fun withCustomStdout(delegateOut: PrintStream, task: () -> Unit) {
     System.setOut(actualOut)
 }
 
-fun winCompatiblePathOf(path: String): Path {
+fun winCompatiblePathOf(path: String): Path =
     if (path.get(2) == ':' && path.get(0) == '/') {
         // Strip leading '/' when dealing with paths on Windows
-        return Paths.get(path.substring(1))
+        Paths.get(path.substring(1))
     } else {
-        return Paths.get(path)
+        Paths.get(path)
     }
-}
 
 fun String.partitionAroundLast(separator: String): Pair<String, String> = lastIndexOf(separator)
     .let { Pair(substring(0, it), substring(it, length)) }
@@ -53,10 +45,8 @@ fun Path.replaceExtensionWith(newExtension: String): Path {
 }
 
 inline fun <T, C : Iterable<T>> C.onEachIndexed(transform: (index: Int, T) -> Unit): C = apply {
-    var i = 0
-    for (element in this) {
+    for ((i, element) in this.withIndex()) {
         transform(i, element)
-        i++
     }
 }
 
@@ -65,26 +55,9 @@ fun <T> noResult(message: String, result: T): T {
     return result
 }
 
-fun <T> noFuture(message: String, contents: T): CompletableFuture<T> = noResult(message, CompletableFuture.completedFuture(contents))
-
 fun <T> emptyResult(message: String): List<T> = noResult(message, emptyList())
 
 fun <T> nullResult(message: String): T? = noResult(message, null)
-
-fun <T> firstNonNull(vararg optionals: () -> T?): T? {
-    for (optional in optionals) {
-        val result = optional()
-        if (result != null) {
-            return result
-        }
-    }
-    return null
-}
-
-fun <T> nonNull(item: T?, errorMsgIfNull: String): T =
-    if (item == null) {
-        throw NullPointerException(errorMsgIfNull)
-    } else item
 
 inline fun <T> tryResolving(what: String, resolver: () -> T?): T? {
     try {
